created method to find prime factor of a given number

+++ src/main/java/geeksforgeeks/mathematics/PrimeNumber.java
@@ -1,7 +1,7 @@
 package main.java.geeksforgeeks.mathematics;
 
 public interface PrimeNumber {
-    public boolean isPrimeUsingNaiveApproach(int n);
-    public boolean isPrimeUsingEfficientApproach(int n);
-    public boolean isPrimeUsingMoreEfficientApproach(int n);
+    public boolean isPrimeUsingNaiveApproach(long n);
+    public boolean isPrimeUsingEfficientApproach(long n);
+    public boolean isPrimeUsingMoreEfficientApproach(long n);
 }
