+++ src/main/java/geeksforgeeks/mathematics/PrimeNumberImpl.java
@@ -2,11 +2,11 @@ package main.java.geeksforgeeks.mathematics;
 
 public class PrimeNumberImpl implements PrimeNumber{
     @Override
-    public boolean isPrimeUsingNaiveApproach(int n) {
+    public boolean isPrimeUsingNaiveApproach(long n) {
         if(n == 1){
             return false;
         }
-        for(int i=2; i<n; i++){
+        for(long i=2; i<n; i++){
             if(n % i ==0){
                 return false;
             }
@@ -15,11 +15,11 @@ public class PrimeNumberImpl implements PrimeNumber{
     }
 
     @Override
-    public boolean isPrimeUsingEfficientApproach(int n) {
+    public boolean isPrimeUsingEfficientApproach(long n) {
         if(n == 1){
             return false;
         }
-        for(int i=2; i*i <=n; i++){
+        for(long i=2; i*i <=n; i++){
             if(n % i == 0){
                 return false;
             }
@@ -28,11 +28,11 @@ public class PrimeNumberImpl implements PrimeNumber{
     }
 
     @Override
-    public boolean isPrimeUsingMoreEfficientApproach(int n) {
+    public boolean isPrimeUsingMoreEfficientApproach(long n) {
         if(n == 1) return false;
         if(n == 2  || n == 3) return true;
         if(n % 2 == 0 || n % 3 == 0) return false;
-        for(int i=5; i*i <= n ; i+=6){
+        for(long i=5; i*i <= n ; i+=6){
             if(n % i ==0 ){
                 return false;
             }
